+++ changelog.md
@@ -1,3 +1,8 @@
 ### 1.1.0
 * Added a button to abort world generation
 * Added compatibility to others mods to include their blocks to the world randomization
+
+### 1.2.0
+* Added an UI to select seed options
+* Set the max seed length to 50 (instead of 20)
+* Fixed few blocks that shouldn't appear in random worlds